Fix COMPARE on maps
It was not comparing the default value.
As a result, the test of the rereading readtable was missing the
bug that the /<default> part of a serialized map was not being read.
Comment out that test case for now.

Also add tests for mutation coverage of COMPOSE (which exposed
the COMPARE bug).