Fix test suite on ABCL, and probably on ECL and CLisp (#42).
For ECL, I implemented `make-char', but then ran into an ECL bug that
kept the test suite from completing
(https://gitlab.com/embeddable-common-lisp/ecl/-/issues/745).

For CLisp, I think the test case was written incorrectly and I have
changed it, but the last CLisp release was in 2010, and when I tried
to build it, it failed.